+++ services/wiki/internal/handler/wiki_cqrs.go
@@ -114,7 +114,7 @@ func (h *WikiCQRSHandler) Update(ctx context.Context, req *pb.UpdateArticleReque
 	}
 
 	// ここの引数はポインターアドレスを渡す必要があるみたいです？
-	article.Update(title, content, visibility, req.IsPinned) // メソッドを呼び出し、既存の値を上書き
+	article.Update(title, content, req.CategoryId, visibility, req.IsPinned) // メソッドを呼び出し、既存の値を上書き
 	err = h.commandRepo.Save(ctx, article)
 	if err != nil {
 		slog.Error("failed to save article", "error", err, "id", req.Id)